SpreadsheetEngineHttps.viewportSelection & anchor from SpreadsheetEngineHateosHandlerSpreadsheetDelta
